Refactor AS 7 adapter distribution  - not dependent on distribution/modules any more  - again use <extension module="org.keycloak.keycloak-adapter-subsystem"/>
