--- distribution/adapters/as7-adapter-zip/assembly.xml
@@ -1,29 +1,0 @@
-<assembly>
-    <id>war-dist</id>
-
-    <formats>
-        <format>zip</format>
-        <format>tar.gz</format>
-    </formats>
-    <includeBaseDirectory>false</includeBaseDirectory>
-
-    <fileSets>
-        <fileSet>
-            <directory>${project.build.directory}/unpacked</directory>
-            <includes>
-                <include>org/bouncycastle/**</include>
-                <include>net/iharder/base64/**</include>
-                <include>org/apache/httpcomponents/**</include>
-                <include>org/keycloak/keycloak-core/**</include>
-                <include>org/keycloak/keycloak-adapter-core/**</include>
-                <include>org/keycloak/keycloak-jboss-adapter-core/**</include>
-                <include>org/keycloak/keycloak-as7-adapter/**</include>
-                <include>org/keycloak/keycloak-as7-subsystem/**</include>
-            </includes>
-            <excludes>
-                <exclude>**/*.war</exclude>
-            </excludes>
-            <outputDirectory>modules</outputDirectory>
-        </fileSet>
-    </fileSets>
-</assembly>
