+++ integration/as7-eap6/as7-subsystem/pom.xml
@@ -21,7 +21,7 @@
         <groupId>org.keycloak</groupId>
         <artifactId>keycloak-parent</artifactId>
         <version>1.3.0.Final-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
+        <relativePath>../../../pom.xml</relativePath>
     </parent>
 
     <artifactId>keycloak-as7-subsystem</artifactId>
