+++ integration/as7-eap6/as7-subsystem/src/main/resources/META-INF/services/org.jboss.as.controller.Extension
