+++ integration/as7-eap6/as7-subsystem/src/main/resources/schema/keycloak_1_1.xsd
