+++ integration/pom.xml
@@ -19,11 +19,10 @@
         <module>servlet-oauth-client</module>
         <module>jboss-adapter-core</module>
         <module>tomcat</module>
-        <module>as7-eap6/adapter</module>
+        <module>as7-eap6</module>
         <module>jetty</module>
         <module>undertow</module>
         <module>wildfly</module>
-        <module>as7-subsystem</module>
         <module>js</module>
         <module>installed</module>
         <module>admin-client</module>
